Implemented full 3d covers.
To test; compile with D3_COVERS enabled in coverflow.h

currently I use a low res version of the full rock band cover for all the covers for testing... need a good generic "NO COVER".

Hover over a cover and hit 1 to flip the cover over.

Coverflow keeps track of which covers are flipped over. (Not saved though).

This change also introduced memory issues... i think it pushed us back over the threshold even with blackbirds change.

We really need to get the bootloader working so memory won't be an issue any more.